+++ tests/integration_tests/functional/test_api.py
@@ -1285,6 +1285,7 @@ def test_get_full_config_after_restoring_snapshot(microvm_factory, uvm_configure
         "smt": True,
         "track_dirty_pages": False,
         "huge_pages": "None",
+        "ksm_mergeable": False,
     }
 
     if cpu_vendor == utils_cpuid.CpuVendor.ARM:
@@ -1430,6 +1431,7 @@ def test_get_full_config(uvm):
         "smt": False,
         "track_dirty_pages": False,
         "huge_pages": "None",
+        "ksm_mergeable": False,
     }
     expected_cfg["cpu-config"] = None
     expected_cfg["boot-source"] = {